+++ src/platforms/android/__tests__/notifications.test.ts
@@ -1,0 +1,126 @@
+import { test } from 'vitest';
+import assert from 'node:assert/strict';
+import { promises as fs } from 'node:fs';
+import os from 'node:os';
+import path from 'node:path';
+import { pushAndroidNotification } from '../notifications.ts';
+import type { DeviceInfo } from '../../../kernel/device.ts';
+
+test('pushAndroidNotification broadcasts action with typed extras', async () => {
+  const tmpDir = await fs.mkdtemp(path.join(os.tmpdir(), 'agent-device-android-push-test-'));
+  const adbPath = path.join(tmpDir, 'adb');
+  const argsLogPath = path.join(tmpDir, 'args.log');
+  await fs.writeFile(
+    adbPath,
+    '#!/bin/sh\nprintf "%s\\n" "$@" > "$AGENT_DEVICE_TEST_ARGS_FILE"\nexit 0\n',
+    'utf8',
+  );
+  await fs.chmod(adbPath, 0o755);
+
+  const previousPath = process.env.PATH;
+  const previousArgsFile = process.env.AGENT_DEVICE_TEST_ARGS_FILE;
+  process.env.PATH = `${tmpDir}${path.delimiter}${previousPath ?? ''}`;
+  process.env.AGENT_DEVICE_TEST_ARGS_FILE = argsLogPath;
+
+  const device: DeviceInfo = {
+    platform: 'android',
+    id: 'emulator-5554',
+    name: 'Pixel',
+    kind: 'emulator',
+    booted: true,
+  };
+
+  try {
+    const result = await pushAndroidNotification(device, 'com.example.app', {
+      action: 'com.example.app.PUSH',
+      extras: {
+        title: 'Hello',
+        unread: 3,
+        promo: true,
+        ratio: 0.5,
+      },
+    });
+    assert.equal(result.action, 'com.example.app.PUSH');
+    assert.equal(result.extrasCount, 4);
+    const args = (await fs.readFile(argsLogPath, 'utf8')).trim().split('\n').filter(Boolean);
+    assert.deepEqual(args, [
+      '-s',
+      'emulator-5554',
+      'shell',
+      'am',
+      'broadcast',
+      '-a',
+      'com.example.app.PUSH',
+      '-p',
+      'com.example.app',
+      '--es',
+      'title',
+      'Hello',
+      '--ei',
+      'unread',
+      '3',
+      '--ez',
+      'promo',
+      'true',
+      '--ef',
+      'ratio',
+      '0.5',
+    ]);
+  } finally {
+    process.env.PATH = previousPath;
+    if (previousArgsFile === undefined) {
+      delete process.env.AGENT_DEVICE_TEST_ARGS_FILE;
+    } else {
+      process.env.AGENT_DEVICE_TEST_ARGS_FILE = previousArgsFile;
+    }
+    await fs.rm(tmpDir, { recursive: true, force: true });
+  }
+});
+
+test('pushAndroidNotification ignores empty extra keys when reporting extrasCount', async () => {
+  const tmpDir = await fs.mkdtemp(
+    path.join(os.tmpdir(), 'agent-device-android-push-empty-key-test-'),
+  );
+  const adbPath = path.join(tmpDir, 'adb');
+  const argsLogPath = path.join(tmpDir, 'args.log');
+  await fs.writeFile(
+    adbPath,
+    '#!/bin/sh\nprintf "%s\\n" "$@" > "$AGENT_DEVICE_TEST_ARGS_FILE"\nexit 0\n',
+    'utf8',
+  );
+  await fs.chmod(adbPath, 0o755);
+
+  const previousPath = process.env.PATH;
+  const previousArgsFile = process.env.AGENT_DEVICE_TEST_ARGS_FILE;
+  process.env.PATH = `${tmpDir}${path.delimiter}${previousPath ?? ''}`;
+  process.env.AGENT_DEVICE_TEST_ARGS_FILE = argsLogPath;
+
+  const device: DeviceInfo = {
+    platform: 'android',
+    id: 'emulator-5554',
+    name: 'Pixel',
+    kind: 'emulator',
+    booted: true,
+  };
+
+  try {
+    const result = await pushAndroidNotification(device, 'com.example.app', {
+      extras: {
+        '': 'ignored',
+        title: 'Welcome',
+      },
+    });
+    assert.equal(result.extrasCount, 1);
+    const args = (await fs.readFile(argsLogPath, 'utf8')).trim();
+    assert.equal(args.includes('\n\n'), false);
+    assert.equal(args.includes('ignored'), false);
+  } finally {
+    process.env.PATH = previousPath;
+    if (previousArgsFile === undefined) {
+      delete process.env.AGENT_DEVICE_TEST_ARGS_FILE;
+    } else {
+      process.env.AGENT_DEVICE_TEST_ARGS_FILE = previousArgsFile;
+    }
+    await fs.rm(tmpDir, { recursive: true, force: true });
+  }
+});
